Set OS_SHARE_API_VERSION envVar for openstackclient
Manilaclient doesn't yet have the API negotiation bits of the OpenStackClient.
This means that we can't create shares unless the env var is exported
every time a shell in the openstackclient pod is requested.
This patch ensures we export the required envVar beforehand, so we don't
have to ask to run the openstack command passing the --os-share-api-version
option every time.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>